Player stats screen (#34)
* Fix for Issue #6

Contains the fix for issue #6

* Added stats screen for Issue #7

Added a simple stats screen for Issue #7

Easily able to add more info to view from the player.

* Fix for Issues #17 and #18

Fixed issues #17 and #18 as well as began fixing issue #5

* Delete Deprecated TODO.txt

The TODO.txt was a legacy file from the pre-GitHub project. Now that the project is keeping track of enhancements and problems through GitHub's issue tracker, it is no longer needed

* Fix for Issue #8

Added a quest view screen that is accessible from the player info menu.

* Fix for Issue #35

Fixed the cleardisplay bug in issue #35.

* Exit Option in Quest View Menu

There was no display for the exit option when in the quest view menu. That is now fixed.

* Compiled Dist

Recomplied the distributable